requirements.txt
python-telegram-bot[job-queue]==20.3
python-dotenv==1.0.0
flask==3.0.2